Refactor ClippingPlane
